BAU-426 Rename DataBlockTabs to DataBlockRenderer
