+++ src/explore-education-statistics-common/src/modules/find-statistics/components/ContentBlockRenderer.tsx
